dbeaver/pro#5267 refactor: auth providers
